Stats/requirements.txt
Django>=4.0,<5.0            # The core Django framework
djangorestframework         # Django REST Framework for building APIs
pymongo                     # MongoDB driver for Python
pandas                      # Pandas for data manipulation and analysis
python-dotenv               # For loading environment variables from .env files
django-cors-headers         # For http headers
gunicorn                    # For deploying Django applications
requests                    # Handle HTTP requests